make sorted_indices and unique_indices top-level
