+++ spec/API_specification/dataframe_api/typing.py
@@ -129,10 +129,18 @@ class Namespace(Protocol):
     def date(self, year: int, month: int, day: int) -> Scalar:
         ...
 
-    def any_horizontal(self, *columns: Column, skip_nulls: bool | Scalar = True) -> Column:
+    def any_horizontal(
+        self,
+        *columns: Column,
+        skip_nulls: bool | Scalar = True,
+    ) -> Column:
         ...
 
-    def all_horizontal(self, *columns: Column, skip_nulls: bool | Scalar = True) -> Column:
+    def all_horizontal(
+        self,
+        *columns: Column,
+        skip_nulls: bool | Scalar = True,
+    ) -> Column:
         ...
